entity and repo
