+++ src/Core/Entity/Entity.js
@@ -1,3 +1,5 @@
+import ORM from "../Database/ORM";
+
 /**
  * @class Entiry
  */
@@ -7,6 +9,47 @@ class Entity {
      */
     constructor() {
 
+    }
+
+    /**
+     * Get all props name
+     * @return {[]}
+     */
+     static getProps() {
+        return ORM.getProps(this.name);
+    }
+
+    /**
+     * get Table Name
+     * @return {string}
+     */
+    static getTable() {
+        return ORM.getTable(this.name);
+    }
+
+    /**
+     * GetP primary key field
+     * @return {string}
+     */
+    static getPrimaryKey() {
+        return ORM.getPrimaryKey(this.name);
+    }
+
+    /**
+     * Get select columns
+     * @return {{}}
+     */
+     static getSelectColumns() {
+        const selectedColumn =[];
+        const colums = this.getProps();
+        for (const [key, value] of Object.entries(colums)) {
+            if (value.isPrimitive) {
+                selectedColumn.push(this.getTable() + '.' + key);
+            } else {
+                selectedColumn.push(this.getTable() + '.' + value.foreignKey);
+            }
+        }
+        return selectedColumn;
     }
 }
